app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["peperoncino <2wua4nlyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mecab-python3 = "^1.0.6"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
fastapi = "^0.89.1"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
